requirements.txt
# Discord Bot Framework
py-cord>=2.6.0
discord-ext-voice-recv>=0.5.2
PyNaCl>=1.5.0

# Voice Recognition & AI
faster-whisper>=1.0.0
torch>=2.0.0
torchaudio>=2.8.0
librosa>=0.10.0
soundfile>=0.12.0
praat-parselmouth>=0.4.0

# AI & LLM Integration
httpx>=0.24.0

# Web Search
duckduckgo-search>=3.9.0

# Music Integration
yt-dlp>=2023.0.0
spotipy>=2.23.0

# Database
sqlcipher3>=0.5.0

# Text-to-Speech
gtts>=2.3.0

# Utilities
python-dotenv>=1.0.0
aiohttp>=3.9.0
numpy>=1.24.0
requests>=2.31.0
